pyproject.toml
[project]
name = "smart-student-agent"
version = "0.1.0"
description = "A smart student agent using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.5.5",
    "python-dotenv>=0.9.9",
    "openai>=1.86.0",
]
